move config class
