+++ src/main/java/de/muenchen/dave/domain/dtos/bearbeiten/BearbeiteZaehlungDTO.java
@@ -1,7 +1,7 @@
 package de.muenchen.dave.domain.dtos.bearbeiten;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import de.muenchen.dave.config.EmptyListSerializer;
+import de.muenchen.dave.configuration.json.EmptyListSerializer;
 import de.muenchen.dave.domain.dtos.PkwEinheitDTO;
 import de.muenchen.dave.domain.enums.Fahrzeug;
 import de.muenchen.dave.validation.BearbeiteZaehlungValid;
